Begin work on nodes-as-crates approach
Currently the graph code generation and Node trait are very limited and
not really practically functional.

Most of the beginning work has been trying to feel out a top-level API
via the `Project` type. The role of the `Project` is to manage the state
of the user's gantz graphs and the cargo workspace that contains
generated packages for each node and their compilation targets.

Currently, an API exists for opening a new/existing project and for
adding nodes. Adding nodes generates packages within the cargo
workspace, though no code generation occurs just yet.

The next step will likely be to get some basic code generation working.

Closes #5.
Closes #6.