Add support for codegen of stateless gantz graph push evaluation
This adds some significant progress to the graph::codegen module -
namely the ability to generate a rust file from a gantz graph. A test
has been added to a new tests directory and the Project API has been
updated accordingly.

The role of serialization has been separated from **Node** into a
**SerdeNode** trait. This should probably moved to the project.rs module
eventually. It seems like it might be useful to separate the project.rs
module into a separate crate at some point as it is quite a high-level
API around the lower-level graph and node code.

The open_project.rs example now creates a graph, generates the src and
updates the root graph node's crate with the new src. The next step will
be attempting to dynamically build, load and execute the generated code.