Add support to allow for optionally typed inputs and outputs
This allows for optionally specifying the full types of the inputs and
outputs of a `Node` during implementation by allowing to specify a full,
freestanding function, rather than only an expression.

The function's arguments and return type will be parsed to produce the
number of inputs and outputs for the node, where the number of arguments
is the number of inputs, and the number of tuple arguments in the output
is the number of outputs (1 if no tuple output type).

Some of this may have to be re-written when addressing a few follow-up
issues including #29, #19, #21 and #22, but I think it's helpful to
break up progress into achievable steps!

Closes #27 and makes #20 much more feasible.