WIP Implementation of `Node` for graph types
This is a WIP implementation of the **Node** trait for graph types of
nodes that also implement the **Node** trait. This work goes toward
allowing nesting of graphs within graphs, the primary form of
abstraction for gantz graph projects.

This works by producing an **Evaluator::Fn** from the
**Node::evaluator** method, where the number of inputs to the generated
function declaration match the number of inlets, and the output type
matches the number of outlets. The implementation is generic over all
graph types, so long as they are able to generate a function body for
the generated declaration mentioned above.

This requirement is specified via the **graph::EvaluatorFnBlock** trait.
An implementation exists for the graph type used within the project
module - it is currently incomplete but specifies the necessary steps
for completion once state handling is added.

Progress on #20.