Early implementation of support for Node State
This adds rough support for node state that persists between calls to
evaluation functions.

Currently, node state is made available in evaluation functions via an
added `&mut [&mut dyn std::any::Any]`. Eventually, this should be
changed to a friendlier `node::States` type or something along these
lines, however this will first require some way to make such a type
available to the generated crate. This might mean splitting this type
into its own crate so that it may be shared between both gantz,
generated crates and user crates. For now, the `Any` slice only requires
`std` and seems to work as a basic prototype.

A simple counter.rs test has been added in which a small graph
containing a counter node is composed, compiled, loaded, evaluated three
times and the result is asserted at the end.

I'm not sure if Rust's unstable ABI will begin causing more issues with
larger types, but for now this seems to be working consistently OK with
primitive types on Linux.

This should make it possible to finish #20.

Closes #19.