`impl Graph for Node` progress
Implements most of the project graph evaluation function generation.
Also refactors the function signature to provide access to the graph's
inlets and outlets as these are necessary for providing access to state
to the nested graph.

Adds a `GraphState` type to use as the graph `Node::state_type`. Need to
implement node crate dependencies in order for this to work, as
generated crates cannot know about the type otherwise.

Going to merge this for now in order to make progress on crate
dependency support, crate refactoring and other items necessary to make
further progress.