Add `graph::full_eval` function generation
This PR adds a step to the codegen process that generates a function for
evaluating the full graph. This function is necessary to call into the
graph when it is nested and retrieve some result. The function is only
generated if the graph contains one or more inlet or outlet nodes, as it
would be impossible to call into the graph and retrieve some result
otherwise.