refactor: Major rewrite targeting Steel rather than Rust
This is a substantial change to gantz, transitioning from generating
Rust dynamic libraries to Steel functions.

Steel is an embeddable scheme written in Rust designed for embedding in
Rust applications.

There are a few motivations for this switch:

- Steel code is structured like an AST, making it easier to target.
- Steel is a lightweight library compared to cargo and rustc.
- Steel "compiles" much faster meaning changes to graphs should be
  reflected much more quickly.
- Steel still allows for specialised native code by registering Rust
  functions in the VM.
- Steel is WASM compatible, opening up the potential for a web version.

The state of this rewrite is roughly equivalent to the previous state of
gantz, though without the top-level `gantz` repository and `Project`
type that were fairly Rust-specific, and mostly managed the generated
rust dynamic libraries and workspaces.

This also includes some other small maintenance tweaks:

- Workspace member crates are now under the `crates/` subdirectory.
- Rust edition updated from 2018 to 2024.
- Make better use of the workspace toml in general.
- Moved tests from `gantz` to `gantz_core`.