feat: Add support for nested graphs to gantz_egui and demo
Allows for adding nested `graph` nodes to an existing graph.

Double-clicking on a nested graph node opens that graph in the graph
scene.

A small representation of the currently selected graph's path from the
root is shown in the bottom left. When viewing a nested graph this
allows for navigating directly to any of the parents.

`inlet` and `outlet` nodes can be added to a subgraph revealing inlets
and outlets on its node when viewed from the parent graph.