feat: Get top-level gantz crate working to gantz_egui demo level
This gets the top-level gantz bevy app working to roughly the same level
as the gantz_egui demo as a starting point.

Currently, most of the logic has been ported directly into the `gantz`
crate, but a lot of this could probably be refactored into `bevy_gantz`
in a more modular manner in the future.

Also adds a DeboundedInputPlugin to `bevy_gantz` to allow for
auto-saving after a short period of inactive user-input.