feat: Add content-addressed commit and head abstractions
This is mostly in anticipation of allowing for viewing multiple heads
with multiple `GraphScene` pane instances.

It should also enable navigating the history of names (similar to
branches), and provide a clean way of handling undo/redo.

Only the `gantz_egui` demo is updated for now - still need to update the
top-level `gantz` crate.